consolidate em_dm_store pacts into single file
