--- src/contractTest/java/uk/gov/hmcts/dm/controller/DocumentStoreProviderTest.java
@@ -1,70 +1,0 @@
-package uk.gov.hmcts.dm.controller;
-
-import au.com.dius.pact.provider.junit5.PactVerificationContext;
-import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
-import au.com.dius.pact.provider.junitsupport.Provider;
-import au.com.dius.pact.provider.junitsupport.State;
-import au.com.dius.pact.provider.junitsupport.loader.PactBrokerConsumerVersionSelectors;
-import au.com.dius.pact.provider.junitsupport.loader.SelectorBuilder;
-import au.com.dius.pact.provider.spring.junit5.MockMvcTestTarget;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.TestTemplate;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.context.annotation.Import;
-import org.springframework.mock.web.MockMultipartFile;
-import uk.gov.hmcts.dm.controller.provider.BaseProviderTest;
-import uk.gov.hmcts.dm.domain.DocumentContentVersion;
-import uk.gov.hmcts.dm.domain.StoredDocument;
-
-import java.nio.charset.StandardCharsets;
-import java.util.Optional;
-import java.util.UUID;
-
-import static org.mockito.Mockito.when;
-
-@Provider("em_dm_store")
-@Import(StoreDocumentControllerTestConfiguration.class)
-public class DocumentStoreProviderTest extends BaseProviderTest {
-
-    private final UUID id = UUID.fromString("5c3c3906-2b51-468e-8cbb-a4002eded075");
-
-    @TestTemplate
-    @ExtendWith(PactVerificationInvocationContextProvider.class)
-    void pactVerificationTestTemplate(PactVerificationContext context) {
-        if (context != null) {
-            context.verifyInteraction();
-        }
-    }
-
-    @PactBrokerConsumerVersionSelectors
-    public static SelectorBuilder consumerVersionSelectors() {
-        return new SelectorBuilder()
-            .matchingBranch()
-            .mainBranch()
-            .deployedOrReleased();
-    }
-
-    @BeforeEach
-    void before(PactVerificationContext context) {
-        System.getProperties().setProperty("pact.verifier.publishResults", "true");
-        if (context != null) {
-            context.setTarget(new MockMvcTestTarget(mockMvc));
-        }
-    }
-
-    @State({"I have existing document"})
-    public void toDeleteDocuments() {
-        DocumentContentVersion documentContentVersion = new DocumentContentVersion(new StoredDocument(),
-            new MockMultipartFile("files",
-                "filename.txt",
-                "text/plain",
-                "hello".getBytes(
-                    StandardCharsets.UTF_8)),
-            "user");
-
-        documentContentVersion.setCreatedBy("userId");
-        when(this.documentContentVersionService.findMostRecentDocumentContentVersionByStoredDocumentId(id))
-            .thenReturn(Optional.of(documentContentVersion));
-
-    }
-}
